pyproject.toml
[tool.poetry]
name = "storageservice"
version = "0.0.1"
description = "This module will store DB data in mongo and image data in minio"
authors = ["TataiQ <faditya.singh@tataiq.com>"]
repository = "https://github.com/./StorageService"
documentation = "https://..github.io/StorageService/"
readme = "README.md"
packages = [
  {include = "storageservice"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
opencv-python = "^4.8.1.78"
minio = "^7.1.17"
pymongo = "^4.5.0"
redis = "^5.0.1"
pyyaml = "^6.0.1"
uvicorn = "^0.23.2"
fastapi = "^0.104.0"
imutils = "^0.5.4"
shared-memory-dict = "^0.7.2"
kafka-python = "^2.0.2"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.981"
pre-commit = "^2.20.0"
tox = "^3.25.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^8.5.10"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
preview = true

[tool.mypy]
files = ["storageservice"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]



